#73 - simplifying CommandIndexTable
