--- src/pages/CommandIndex/CommandIndexTable/CommandIndexTable.tsx
@@ -1,37 +1,0 @@
-import { Box, Checkbox, FormControlLabel } from '@mui/material'
-import Breadcrumbs from 'components/Breadcrumbs'
-import { Table } from 'components/Table'
-import {
-  useCommandIndexTableColums,
-  useCommands,
-} from 'pages/CommandIndex/CommandIndexTable'
-
-const CommandIndexTable = () => {
-  const {
-    commands,
-    namespace,
-    systemName,
-    version,
-    includeHidden,
-    hiddenOnChange,
-  } = useCommands()
-  const columns = useCommandIndexTableColums()
-  const breadcrumbs = [namespace, systemName, version]
-    .filter((x) => !!x)
-    .map((x) => String(x))
-
-  return (
-    <Table tableName="Commands" data={commands} columns={columns}>
-      <Box mb={2}>
-        <Breadcrumbs breadcrumbs={breadcrumbs} />
-        <FormControlLabel
-          label="Include hidden"
-          control={
-            <Checkbox checked={includeHidden} onChange={hiddenOnChange} />
-          }
-        />
-      </Box>
-    </Table>
-  )
-}
-export { CommandIndexTable }
