--- src/pages/CommandIndex/CommandIndexTable/index.ts
@@ -1,2 +1,0 @@
-export * from './CommandIndexTable'
-export * from './data'
