+++ src/pages/CommandIndex/index.ts
@@ -1,2 +1,3 @@
 export * from './CommandIndex'
+export * from './CommandIndexTable'
 export * from './MakeItHappenButton'
